fix: separate the utils into independent modules
