+++ package.json
@@ -39,10 +39,10 @@
     "clean": "shx rm -rf build ./node_modules/zeromq/build ./node_modules/zeromq/staging",
     "lint.eslint": "eslint . --cache --cache-location ./.cache/.eslintcache \"./**/*.{js,ts,mjs,mts,cjs,cts,json,yaml,yml}\" --fix",
     "lint.biome": "biome check --write --unsafe .",
-    "lint": "turbo run lint.eslint lint.biome",
+    "lint": "turbo run lint.eslint lint.biome format.prettier format.biome",
     "test.lint.eslint": "eslint . --cache --cache-location ./.cache/.eslintcache \"./**/*.{js,ts,mjs,mts,cjs,cts,json,yaml,yml}\"",
     "test.lint.biome": "biome check .",
-    "test.lint": "turbo run test.lint.eslint test.lint.biome",
+    "test.lint": "turbo run test.lint.eslint test.lint.biome test.lint.prettier test.format.biome",
     "format.prettier": "prettier -l --write --cache --cache-location ./.cache/.prettiercache \"./**/*.{yaml,yml,md}\"",
     "format.biome": "biome format --write .",
     "format": "turbo run format.prettier format.biome",
