+++ src/lib.ts
@@ -1,4 +1,4 @@
 export * from "./build.js"
 export * from "./config-types.d"
-export * from "./logger.js"
+export * from "./utils/logger.js"
 export * from "./loader.js"
