+++ src/urlRegistry.ts
@@ -2,7 +2,7 @@ import os from "os"
 import gte from "semver/functions/gte"
 import lt from "semver/functions/lt"
 import type { BuildConfiguration } from "./config-types.d"
-import { getEnvVar } from "./util.js"
+import { getEnvVar } from "./utils/env.js"
 
 const NODE_MIRROR = getEnvVar("NVM_NODEJS_ORG_MIRROR") ?? "https://nodejs.org/dist"
 const IOJS_MIRROR = getEnvVar("NVM_IOJS_ORG_MIRROR") ?? "https://iojs.org/dist"
