+++ src/utils/env.ts
@@ -1,0 +1,13 @@
+/**
+ * Get an environment variable.
+ *
+ * @param name - The name of the environment variable.
+ * @returns The value of the environment variable or undefined if it is not set.
+ */
+export function getEnvVar(name: string) {
+  const value = process.env[name]
+  if (typeof value === "string" && value.length > 0) {
+    return value
+  }
+  return undefined
+}
